requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
openai-whisper
gtts
python-multipart